requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
anyio=3.5.0=py38hecd8cb5_0
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=20.1.0=py38h9ed2024_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=py38hecd8cb5_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py38h67323c0_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotli-python=1.0.9=py38he9d5cce_7
ca-certificates=2023.12.12=hecd8cb5_0
certifi=2024.2.2=py38hecd8cb5_0
cffi=1.16.0=py38h6c40b1e_0
comm=0.1.2=py38hecd8cb5_0
contourpy=1.0.5=py38haf03e11_0
cryptography=41.0.7=py38h30e54ef_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py38hcec6c5f_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38hecd8cb5_0
exceptiongroup=1.0.4=py38hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
geographiclib=1.52=pyhd8ed1ab_0
geopy=2.4.1=pyhd8ed1ab_0
giflib=5.2.1=h6c40b1e_3
h11=0.12.0=pyhd3eb1b0_0
idna=3.4=py38hecd8cb5_0
importlib_resources=6.1.1=py38hecd8cb5_1
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.25.0=py38h20db666_0
ipython=8.12.2=py38hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py38hecd8cb5_1
jinja2=3.1.2=py38hecd8cb5_0
jpeg=9e=h6c40b1e_1
jsonschema=4.19.2=py38hecd8cb5_0
jsonschema-specifications=2023.7.1=py38hecd8cb5_0
jupyter_client=7.4.9=py38hecd8cb5_0
jupyter_core=5.5.0=py38hecd8cb5_0
jupyter_server=1.13.5=pyhd3eb1b0_0
kiwisolver=1.4.4=py38hcec6c5f_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libffi=3.4.4=hecd8cb5_0
libpng=1.6.39=h6c40b1e_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.5.1=hcec6c5f_0
libwebp=1.3.2=hf6ce154_0
libwebp-base=1.3.2=h6c40b1e_0
lz4-c=1.9.4=hcec6c5f_0
markupsafe=2.1.3=py38h6c40b1e_0
matplotlib=3.7.1=py38hecd8cb5_1
matplotlib-base=3.7.1=py38hda11e5a_1
matplotlib-inline=0.1.6=py38hecd8cb5_0
mistune=0.8.4=py38h1de35cc_1001
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.1=py38h4ab4a9b_0
mkl_random=1.2.2=py38hb2f4e1b_0
munkres=1.1.4=py_0
nbclassic=1.0.0=py38hecd8cb5_0
nbconvert=5.4.1=py_2
nbformat=5.9.2=py38hecd8cb5_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=py38hecd8cb5_0
notebook=6.5.4=py38hecd8cb5_1
notebook-shim=0.2.3=py38hecd8cb5_0
numexpr=2.8.4=py38he696674_0
numpy=1.24.3=py38he696674_0
numpy-base=1.24.3=py38h9cd3388_0
openjpeg=2.4.0=h66ea3da_0
openssl=3.0.13=hca72f7f_0
outcome=1.1.0=pyhd3eb1b0_0
packaging=23.1=py38hecd8cb5_0
pandas=1.5.3=py38h07fba90_0
pandoc=2.12=hecd8cb5_3
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.0.1=py38h7d39338_0
pip=23.3.1=py38hecd8cb5_0
pkgutil-resolve-name=1.3.10=py38hecd8cb5_1
platformdirs=3.10.0=py38hecd8cb5_0
prometheus_client=0.14.1=py38hecd8cb5_0
prompt-toolkit=3.0.43=py38hecd8cb5_0
psutil=5.9.0=py38hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py38hecd8cb5_1
pyopenssl=23.2.0=py38hecd8cb5_0
pyparsing=3.0.9=py38hecd8cb5_0
pysocks=1.7.1=py38_1
python=3.8.18=h5ee71fb_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38hecd8cb5_0
pytz=2023.3.post1=py38hecd8cb5_0
pyzmq=23.2.0=py38he9d5cce_0
readline=8.2=hca72f7f_0
referencing=0.30.2=py38hecd8cb5_0
rpds-py=0.10.6=py38hf2ad997_0
selenium=4.9.0=pyhd8ed1ab_0
send2trash=1.8.2=py38hecd8cb5_0
setuptools=68.2.2=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py38hecd8cb5_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.17.1=py38hecd8cb5_0
testpath=0.6.0=py38hecd8cb5_0
tk=8.6.12=h5d9f67b_0
tornado=6.3.3=py38h6c40b1e_0
traitlets=5.7.1=py38hecd8cb5_0
trio=0.24.0=py38hecd8cb5_0
trio-websocket=0.11.1=pyhd8ed1ab_0
typing_extensions=4.9.0=py38hecd8cb5_1
urllib3=1.26.18=py38hecd8cb5_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38hecd8cb5_4
wheel=0.41.2=py38hecd8cb5_0
wsproto=1.2.0=pyhd8ed1ab_0
xz=5.4.5=h6c40b1e_0
zeromq=4.3.5=hcec6c5f_0
zipp=3.17.0=py38hecd8cb5_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
annotated-types=0.6.0=py39hecd8cb5_0
anyio=4.2.0=py39hecd8cb5_0
blas=1.0=mkl
bottleneck=1.3.7=py39h7b7cdfe_0
ca-certificates=2023.12.12=hecd8cb5_0
certifi=2024.2.2=py39hecd8cb5_0
cffi=1.16.0=py39h6c40b1e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39hecd8cb5_0
cryptography=42.0.5=py39h30e54ef_0
ecdsa=0.17.0=pyhd3eb1b0_0
exceptiongroup=1.2.0=py39hecd8cb5_0
fastapi=0.103.0=py39hecd8cb5_0
geographiclib=2.0=py39hecd8cb5_0
geopy=2.4.1=pyhd8ed1ab_1
h11=0.14.0=py39hecd8cb5_0
idna=3.4=py39hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43548
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py39h6c40b1e_1
mkl_fft=1.3.8=py39h6c40b1e_0
mkl_random=1.2.4=py39ha357a0b_0
ncurses=6.4=hcec6c5f_0
numexpr=2.8.7=py39h827a554_0
numpy=1.26.4=py39h827a554_0
numpy-base=1.26.4=py39ha186be2_0
openssl=3.0.13=hca72f7f_0
pandas=2.2.1=py39hcec6c5f_0
pip=23.3.1=py39hecd8cb5_0
pyasn1=0.4.8=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.5.3=py39hecd8cb5_0
pydantic-core=2.14.6=py39hf2ad997_0
python=3.9.18=h5ee71fb_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-jose=3.3.0=pyh6c4a22f_1
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2023.3.post1=py39hecd8cb5_0
readline=8.2=hca72f7f_0
requests=2.31.0=py39hecd8cb5_1
rsa=4.7.2=pyhd3eb1b0_1
setuptools=68.2.2=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py39hecd8cb5_0
sqlite=3.41.2=h6c40b1e_0
starlette=0.27.0=py39hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tk=8.6.12=h5d9f67b_0
typing-extensions=4.9.0=py39hecd8cb5_1
typing_extensions=4.9.0=py39hecd8cb5_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py39hecd8cb5_0
uvicorn=0.28.0=py39h6e9494a_0
wheel=0.41.2=py39hecd8cb5_0
xz=5.4.6=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
